creates league stats module
