+++ lib/team_stats.rb
@@ -15,49 +15,10 @@ module TeamStats
     team_hash[id] = find_team(id).provide_info
   end
 
-  def lowest_scoring_visitor
-    lowest_scoring_away_team = teams.min_by do |team|
-      if team.games_played_as_visitor(games) != 0
-        team.total_away_points(games).to_f / team.games_played_as_visitor(games)
-      else
-        100
-      end
-    end
-    lowest_scoring_away_team.team_name
-  end
+  
 
-  def lowest_scoring_home_team
-    lowest_scoring_home_team = teams.min_by do |team|
-      if team.games_played_as_home_team(games) != 0
-        team.total_home_points(games).to_f / team.games_played_as_home_team(games)
-      else
-        100
-      end
-    end
-    lowest_scoring_home_team.team_name
-  end
 
-  def highest_scoring_home_team
-    highest_scoring_home_team = teams.max_by do |team|
-      if team.games_played_as_home_team(games) != 0
-        team.total_home_points(games).to_f / team.games_played_as_home_team(games)
-      else
-        0
-      end
-    end
-    highest_scoring_home_team.team_name
-  end
 
-  def highest_scoring_visitor
-    highest_scoring_away_team = teams.max_by do |team|
-      if team.games_played_as_visitor(games) != 0
-        team.total_away_points(games).to_f / team.games_played_as_visitor(games)
-      else
-        0
-      end
-    end
-    highest_scoring_away_team.team_name
-  end
 
   def winningest_team
     team_with_highest_win_percentage = @teams.max_by do |team|
